Build batchProcessor with typing & concurrency
batchProcessor was meant to do 3 things:
1. Hold logic for processing one task
2. Split a list of tasks into batches
3. Process a batch

Single-task processing logic will move elsewhere in the future - it will
need to be different depending on whether a user is running this locally
from the command line or the web server is running this, processing for
many users.

For processing a batch, I'd initially wanted to use node workers but it
feels like overkill right now. `Promise.all()` will suffice.

The typing sucks but that will be improved in the future.

Cleared `processInfo.ts` for now. In my original design, that file
looked like a good home for some logic but now makes no sense.